+++ xmake/modules/devel/git/clone.lua
@@ -1,0 +1,74 @@
+--!The Make-like Build Utility based on Lua
+--
+-- Licensed to the Apache Software Foundation (ASF) under one
+-- or more contributor license agreements.  See the NOTICE file
+-- distributed with this work for additional information
+-- regarding copyright ownership.  The ASF licenses this file
+-- to you under the Apache License, Version 2.0 (the
+-- "License"); you may not use this file except in compliance
+-- with the License.  You may obtain a copy of the License at
+--
+--     http://www.apache.org/licenses/LICENSE-2.0
+--
+-- Unless required by applicable law or agreed to in writing, software
+-- distributed under the License is distributed on an "AS IS" BASIS,
+-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+-- See the License for the specific language governing permissions and
+-- limitations under the License.
+-- 
+-- Copyright (C) 2015 - 2017, TBOOX Open Source Group.
+--
+-- @author      ruki
+-- @file        clone.lua
+--
+
+-- imports
+import("core.base.option")
+import("detect.tool.find_git")
+
+-- clone url
+--
+-- @param url   the git url
+-- @param opt   the argument options
+--
+-- @code
+--
+-- import("devel.git")
+-- 
+-- git.clone("git@github.com:tboox/xmake.git")
+-- git.clone("git@github.com:tboox/xmake.git", {depth = 1, branch = "master", outputdir = "/tmp/xmake"})
+--
+-- @endcode
+--
+function main(url, opt)
+
+    -- find git
+    local program = find_git()
+    if not program then
+        return 
+    end
+
+    -- init argv
+    local argv = {"clone", url}
+
+    -- set branch
+    opt = opt or {}
+    if opt.branch then
+        table.insert(argv, "-b")
+        table.insert(argv, opt.branch)
+    end
+
+    -- set depth
+    if opt.depth then
+        table.insert(argv, "--depth")
+        table.insert(argv, ifelse(type(opt.depth) == "number", tostring(opt.depth), opt.depth))
+    end
+
+    -- set outputdir
+    if opt.outputdir then
+        table.insert(argv, path.translate(opt.outputdir))
+    end
+
+    -- clone it
+    os.vrunv(program, argv)
+end